sdk/python-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ezauth-sdk"
version = "0.1.0"
description = "EZAuth Python server SDK"
requires-python = ">=3.11"
dependencies = [
    "python-jose[cryptography]>=3.3",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ezauth_sdk"]
